swap enhancement

+++ swap-setup.zdb/apply.sh
@@ -2,6 +2,10 @@
 source params.sh
 test ! -z "$size_mb"
 test ! -z "$path"
+if test -f "$path"; then
+  echo "swap file already exist. skipping"
+  exit 0
+fi
 
 dd if=/dev/zero of="$path" bs=1M count="$size_mb" status=progress
 chmod 0600 "$path"
